refactor: rename gotodir to cd
